backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0

# HTTP Client
httpx==0.25.2

# Redis (for sessions and caching)
redis==5.0.1
aioredis==2.0.1

# Rate Limiting
slowapi==0.1.9

# Environment and Configuration
python-dotenv==1.0.0
pydantic[email]==2.5.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
